+++ common/src/utils/common/threadqueue.h
@@ -181,7 +181,7 @@ private:
     {
         QMutexLocker mutex(&m_cs);
 
-        int oldSize = m_buffer.size();
+        int oldSize = (int) m_buffer.size();
         m_buffer.resize(newSize);
 
         if (m_headIndex > 0 && m_bufferLen > 0 && newSize > oldSize)
