+++ common/src/utils/network/http/httpstreamreader.cpp
@@ -304,7 +304,7 @@ namespace nx_http
                     }
 
                     const size_t bytesToCopy = std::min<>( m_currentChunkSize - m_currentChunkBytesRead, maxOffset - currentOffset );
-                    m_msgBodyBuffer.append( data.data()+currentOffset, bytesToCopy );
+                    m_msgBodyBuffer.append( data.data()+currentOffset, (int) bytesToCopy );
                     m_messageBodyBytesRead += bytesToCopy;
                     m_currentChunkBytesRead += bytesToCopy;
                     currentOffset += bytesToCopy;
@@ -369,7 +369,7 @@ namespace nx_http
         const size_t bytesToCopy = m_contentLength > 0
             ? std::min<size_t>( count, m_contentLength-m_messageBodyBytesRead )
             : count;    //Content-Length is unknown
-        m_msgBodyBuffer.append( data.data() + offset, bytesToCopy );
+        m_msgBodyBuffer.append( data.data() + offset, (int) bytesToCopy );
         m_messageBodyBytesRead += bytesToCopy;
         return bytesToCopy;
     }
